refactor(plugin-dir): improve class documentation and enforce singleton pattern
